Split class permissions tests into new file
